Invocation Persistence Halfway

+++ src/vs/workbench/contrib/debug/browser/swarmClasses/services/eventService.ts
@@ -19,7 +19,7 @@ export class EventService {
 			const query = `
 			mutation eventCreate(
 				$sessionId: Long,
-				$eventLineNumber: Long,
+				$eventLineNumber: Int,
 				$eventKind: String,
 				$methodId: Long
 			){
